+++ compile/index.js
@@ -1,22 +1,22 @@
-const { _myNewPackage } = require('./mnp')
+const { _amazingPackage } = require('./amazing-package')
 
 /**
- * {{ description }}
- * @param {!_myNewPackage.Config} config Options for the program.
+ * Hello World
+ * @param {!_amazingPackage.Config} config Options for the program.
  * @param {boolean} [config.shouldRun=true] A boolean option. Default `true`.
  * @param {string} [config.text] A text to return.
  * @return {Promise<string>}
  */
-function myNewPackage(config) {
-  return _myNewPackage(config)
+function amazingPackage(config) {
+  return _amazingPackage(config)
 }
 
-module.exports = myNewPackage
+module.exports = amazingPackage
 
 /* typal types/index.xml namespace */
 /**
- * @typedef {_myNewPackage.Config} Config `＠record` Options for the program.
- * @typedef {Object} _myNewPackage.Config `＠record` Options for the program.
+ * @typedef {_amazingPackage.Config} Config `＠record` Options for the program.
+ * @typedef {Object} _amazingPackage.Config `＠record` Options for the program.
  * @prop {boolean} [shouldRun=true] A boolean option. Default `true`.
  * @prop {string} [text] A text to return.
  */