+++ src/bin/amazing-package.js
@@ -1,4 +1,21 @@
-/* license-copyright */
+/*!
+ * amazing-package: Hello World
+ *
+ * Copyright (C) 2020  Art Deco Code Limited
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
 
 import { _help, _init, _output, _version, _input, argsConfig } from './get-args'
 import { reduceUsage } from 'argufy'
@@ -6,13 +23,13 @@ import usually from 'usually'
 import { readFileSync, writeFileSync } from 'fs'
 import { c } from 'erte'
 import Init from './commands/init'
-import myNewPackage from '../'
+import amazingPackage from '../'
 
 if (_help) {
   const usage = usually({
-    description: '{{ description }}',
-    example: 'mnp example.txt -o out.txt',
-    line: 'mnp input [-o output] [-ihv]',
+    description: 'Hello World',
+    example: 'amazing-package example.txt -o out.txt',
+    line: 'amazing-package input [-o output] [-ihv]',
     usage: reduceUsage(argsConfig),
   })
   console.log(usage)
@@ -27,7 +44,7 @@ if (_help) {
     if (_init) return await Init()
     if (!_input) throw new Error('Please pass an input file.')
     const content = /** @type {string} */ (readFileSync(_input, 'utf8'))
-    const output = await myNewPackage({
+    const output = await amazingPackage({
       shouldRun: true,
       text: content,
     })