+++ src/index.js
@@ -1,19 +1,19 @@
 import { c } from 'erte'
 
 /**
- * @type {_myNewPackage.myNewPackage}
+ * @type {_amazingPackage.amazingPackage}
  */
-export default async function myNewPackage(config = {}) {
+export default async function amazingPackage(config = {}) {
   const {
     shouldRun = true,
     text = '',
   } = config
   if (!shouldRun) return ''
-  console.log('my-new-package called with %s', c(text, 'yellow'))
+  console.log('amazing-package called with %s', c(text, 'yellow'))
   return text
 }
 
 /**
  * @suppress {nonStandardJsDocs}
- * @typedef {import('../types').myNewPackage} _myNewPackage.myNewPackage
+ * @typedef {import('../types').amazingPackage} _amazingPackage.amazingPackage
  */