+++ test/context/index.js
@@ -2,7 +2,7 @@ import { join } from 'path'
 import { debuglog } from 'util'
 import { readFileSync } from 'fs'
 
-const LOG = debuglog('my-new-package')
+const LOG = debuglog('amazing-package')
 
 /**
  * A testing context for the package.
@@ -49,10 +49,10 @@ export default class Context {
 let BIN = 'src/BIN'
 if (process.env.ALAMODE_ENV == 'test-build') {
   console.log('Testing build bin...')
-  BIN = 'build/bin/mnp'
+  BIN = 'build/bin/amazing-package'
 } else if (process.env.ALAMODE_ENV == 'test-compile') {
   console.log('Testing compile bin...')
-  BIN = 'compile/bin/mnp'
+  BIN = 'compile/bin/amazing-package'
 }
 
 /** @typedef {Object<string, Test & TestSuite4>} TestSuite */