+++ test/spec/default.js
@@ -1,19 +1,19 @@
 import { equal, ok } from '@zoroaster/assert'
 import Context from '../context'
-import myNewPackage from '../../src'
+import amazingPackage from '../../src'
 
 /** @type {TestSuite} */
 const T = {
   context: Context,
   'is a function'() {
-    equal(typeof myNewPackage, 'function')
+    equal(typeof amazingPackage, 'function')
   },
   async 'calls package without error'() {
-    await myNewPackage()
+    await amazingPackage()
   },
   async 'gets a link to the fixture'({ fixture }) {
     const text = fixture`text.txt`
-    const res = await myNewPackage({
+    const res = await amazingPackage({
       text,
     })
     ok(res, text)