+++ types/index.js
@@ -2,10 +2,10 @@ export {}
 
 /* typal types/api.xml namespace */
 /**
- * @typedef {_myNewPackage.myNewPackage} myNewPackage {{ description }}
- * @typedef {(config: !_myNewPackage.Config) => string} _myNewPackage.myNewPackage {{ description }}
+ * @typedef {_amazingPackage.amazingPackage} amazingPackage Hello World
+ * @typedef {(config: !_amazingPackage.Config) => string} _amazingPackage.amazingPackage Hello World
  */
 
 /**
- * @typedef {import('..').Config} _myNewPackage.Config
+ * @typedef {import('..').Config} _amazingPackage.Config
  */